renaming test file for consistency
